# openapi-schema.yaml
openapi: 3.0.3
info:
  title: Training Manager API
  version: 1.0.0
  description: API pour la gestion d entrainements
paths:
...
  /api/v1/rounds/{id}/clone-exercise/:
    post:
      operationId: rounds_clone_exercise_create
      description: Clone an Exercise and attach the copy to this Round.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this Round.
        required: true
      tags:
      - rounds
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CloneExerciseRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CloneExerciseRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CloneExerciseRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exercise'
          description: ''
        '400':
          description: No response body
        '404':
          description: No response body
  /api/v1/rounds/{id}/exercises/reorder/:
    post:
      operationId: rounds_exercises_reorder_create
      description: Atomically reorder the Exercises attached to this Round. `exercise_ids`
        must contain exactly the IDs of the Exercises currently attached, in the desired
        final order. Exercise.order is set to 1..N matching list position, in a single
        transaction.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this Round.
        required: true
      tags:
      - rounds
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderExercisesRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ReorderExercisesRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ReorderExercisesRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '204':
          description: Exercises reordered
        '400':
          description: 'Validation error. `body.code` is one of: `empty_list`, `duplicate_id`,
            `scope_mismatch`, `incomplete_reorder`.'
        '403':
          description: Not authorized to mutate this round
...
components:
  schemas:
...
    CloneExerciseRequest:
      type: object
      properties:
        exercise_id:
          type: integer
      required:
      - exercise_id
...
    EnergySegment:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        abv:
          type: string
          maxLength: 10
        description:
          type: string
          nullable: true
          maxLength: 200
        energy_system:
          allOf:
          - $ref: '#/components/schemas/EnergySystem'
          readOnly: true
        energy_system_id:
          type: integer
          writeOnly: true
          nullable: true
        is_active:
          type: boolean
          readOnly: true
      required:
      - abv
      - energy_system
      - id
      - is_active
...
    EnergySystem:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          readOnly: true
        is_active:
          type: boolean
          readOnly: true
      required:
      - id
      - is_active
      - name
...
    Exercise:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        order:
          type: integer
          maximum: 9223372036854775807
          minimum: -9223372036854775808
          format: int64
        repetition:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        distance:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        notes:
          type: string
          maxLength: 200
        t_start:
          type: string
          nullable: true
          description: MM:SS format, e.g. 1:30.
          pattern: ^\d{1,3}:[0-5]\d$
        t_break:
          type: string
          nullable: true
          description: MM:SS format, e.g. 1:30.
          pattern: ^\d{1,3}:[0-5]\d$
        modality:
          allOf:
          - $ref: '#/components/schemas/Modality'
          readOnly: true
        modality_id:
          type: integer
          writeOnly: true
          nullable: true
        energysegment:
          allOf:
          - $ref: '#/components/schemas/EnergySegment'
          readOnly: true
        energysegment_id:
          type: integer
          writeOnly: true
          nullable: true
        round_id:
          type: integer
          writeOnly: true
          description: Optional. If provided on POST, the newly created Exercise is
            atomically attached to this Round. The request user must manage at least
            one team of at least one Event linked to this Round (library rounds with
            no events are accepted as-is). Ignored on PATCH/PUT.
        language:
          $ref: '#/components/schemas/LanguageEnum'
        usage_count:
          type: integer
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - energysegment
      - id
      - language
      - modality
      - updated_at
      - usage_count
...
    LanguageEnum:
      enum:
      - fr
      - nl
      - en
      - it
      - es
      type: string
      description: |-
        * `fr` - Français
        * `nl` - Nederlands
        * `en` - English
        * `it` - Italiano
        * `es` - Español
...
    Modality:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          readOnly: true
        sport:
          allOf:
          - $ref: '#/components/schemas/Sport'
          readOnly: true
        is_active:
          type: boolean
          readOnly: true
      required:
      - id
      - is_active
      - name
      - sport
...
    ReorderExercisesRequest:
      type: object
      description: |-
        Body for POST /rounds/{id}/exercises/reorder/.

        `exercise_ids` must contain exactly the IDs of the Exercises currently
        attached to the Round in the desired final order. The view enforces
        scope/completeness in addition to this serializer's syntactic check.
      properties:
        exercise_ids:
          type: array
          items:
            type: integer
            minimum: 1
      required:
      - exercise_ids
...
    Sport:
      type: object
      description: |-
        Public read serializer: 'name' is auto-localised by modeltranslation.
        energy_systems exposed as PK list (frontend joins via /energy-systems/).
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          readOnly: true
        slug:
          type: string
          readOnly: true
          pattern: ^[-a-zA-Z0-9_]+$
        is_active:
          type: boolean
          readOnly: true
        energy_systems:
          type: array
          items:
            type: integer
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - energy_systems
      - id
      - is_active
      - name
      - slug
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT